% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wat_catcher.R
\docType{class}
\name{WatCatcher-class}
\alias{WatCatcher}
\alias{WatCatcher-class}
\title{A reference class for posting wats to wattle}
\description{
A reference class for posting wats to wattle
}
\section{Fields}{

\describe{
\item{\code{exception}}{The simpleError object for the raised exception}
}}
\section{Methods}{

\describe{
\item{\code{catch()}}{Post the wat to the specified wattle host}

\item{\code{check_for_params()}}{Validate that all necessary params are present}

\item{\code{full_host()}}{Create the full path to post the wat to}

\item{\code{initialize(e)}}{Assign the exception field and validate that all necessary wattle params are present}

\item{\code{wat_payload()}}{Build the post payload to be sent to wattle}
}}

